This updates the Cuss Words
Also, i made edits to the function but it disappeared(ALL OF IT)

+++ bot.js
@@ -25,7 +25,7 @@ var contains = function(a, b) {
   }
 };
 
-var badWords = " fuck, shit, ass, yoy, cock, dick, sex, porn, fucker, mother fucker, bitch, asshole, tit, vagina, pussy".split(",");
+var badWords = "fuck, shit, ass, cock, dick, sex, porn, fucker, mother fucker, bitch, asshole, tit, vagina, pussy, Hell, FUCK, Fuck, Fucker, Shit, Ass, Dick, Cock, Sex, Porn".split(", ");
 
 
 
@@ -70,23 +70,23 @@ if (msg.content.startsWith(prefix + "eval")) {
     })
       }
     }
- /*   if(said.startsWith(prefix + "tag")) {
+  if(said.startsWith(prefix + "tag")) {
       var say;
       var toTag = msg.content.slice(prefix.length+4, msg.length);
    
         switch(toTag) {
-    case rolerewards:
-        say = "**WHAT DO I GET FOR CHATTING HERE**\nTake a look at the role rewards! https://shadowka.gitbooks.io/clyde/content/#rolerewards"
-        break;
-    case clydeadmins:
-        say = "The Clyde Admins are Aqil#4788 and Shadow™#8337."
-        break;
-    default:
-        say = "`ERROR`\nTag unavailable."
-            
-}
+          case rolerewards:
+              say = "**WHAT DO I GET FOR CHATTING HERE**\nTake a look at the role rewards! https://shadowka.gitbooks.io/clyde/content/#rolerewards"
+              break;
+          case clydeadmins:
+              say = "The Clyde Admins are Aqil#4788 and Shadow™#8337."
+              break;
+          default:
+              say = "`ERROR`\nTag unavailable."
+
+        }
       msg.channel.send(say)
-    }*/
+    }
     if(msg.content === prefix + "suggest party-instructions") {
       msg.delete()
       msg.channel.send(":scream: _A new channel just appeared for me! What is it???_\nThis is the place to submit your suggestions for parties in the next AqilAcademy election.\n\n**Okay, I just had this great idea for a party! How do I submit it?**\nIt's simple! Just use the `.suggestparty` command anywhere on this server. It will appear here and can be voted on by other members.\n```Example: .suggestparty The Party of ™Ness```\n**I see some great suggestions here, how do I support them?** OR **I don't like this suggestion, how do I downvote it?**\nReact with either <:upvote:361166616081203200> or <:downvote:361166591943114752>.")
